pyproject.toml
[project]
name = "tata"
version = "0.1.0"
description = "Tata - GlobalConnect's Talent Acquisition Team Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tata"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
markers = [
    "integration: marks tests as integration tests requiring real OpenAI API (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["src/tata"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "Protocol",
    "@abstractmethod",
]
